chore: Use yarn and support production builds
This commit converts the project to use Yarn as its package manager. The immediate goal is to improve reliability and speed, and share common dependencies. Future commits may choose to lean more into Yarn's features (such as workspace dependencies or Plug'n'Play installations).

Most of the changes are straightforward. The documentation was updated: you can use all previous commands, except by replacing the `npm` prefix by `npm`.

This commit fixes some smaller issues with dependencies. The `react@*` specification is restricted to match React 17, and the `@mui/lab` package was pinned exactly to the version `5.0.0-alpha.75`.

This commit also adds a `build:production` to generate a full build ahead of time. As part of this change, the server code was moved to a subdirectory.

Finally, this also adds the top-level commands `lint` and `format` to run ESLint in "check" or "fix" mode.