fix: skip logging common oauth token claim errors
This commit adds a check for the Eternaltwin error code when claiming an Oauth token fails.
It then skips sending the error if the error is due to a reused token, this usually appears as an expired token or a token with an invalid format.

This also updates Eternaltwin dependencies to their latest version.